refactor: MyPageResponse 패키지 수정
