+++ backend/src/main/java/edonymyeon/backend/member/application/dto/response/MyPageResponse.java
@@ -1,4 +1,4 @@
-package edonymyeon.backend.member.application.dto;
+package edonymyeon.backend.member.application.dto.response;
 
 //todo 프로필 사진
 public record MyPageResponse(Long memberId, String nickname) {
